DRY paymentRow* and related data
