+++ frontend/views/containers/payments/PaymentsList.vue
@@ -3,9 +3,8 @@ table.table.table-in-card.c-payments
   thead
     tr
       th {{ titles.one }}
-      th.c-payments-amount {{ titles.two }}
-      th.c-payments-date {{ titles.three }}
-
+      th.c-th-amount {{ titles.two }}
+      th.c-th-actions {{ titles.three }}
   tbody
     component(
       v-for='(payment, index) in paymentsList'
@@ -16,13 +15,11 @@ table.table.table-in-card.c-payments
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
 import AvatarUser from '@components/AvatarUser.vue'
 import Tooltip from '@components/Tooltip.vue'
-import PaymentRowTodo from '@containers/payments/PaymentRowTodo.vue'
-import PaymentRowSent from '@containers/payments/PaymentRowSent.vue'
-import PaymentRowReceived from '@containers/payments/PaymentRowReceived.vue'
-import currencies from '@view-utils/currencies.js'
+import PaymentRowTodo from './payment-row/PaymentRowTodo.vue'
+import PaymentRowSent from './payment-row/PaymentRowSent.vue'
+import PaymentRowReceived from './payment-row/PaymentRowReceived.vue'
 
 export default {
   name: 'PaymentsList',
@@ -48,34 +45,7 @@ export default {
       required: true
     }
   },
-  data () {
-    return {
-      // Temp
-      tableChecked: false
-    }
-  },
-  computed: {
-    ...mapGetters([
-      'currentGroupState',
-      'paymentTotalFromUserToUser',
-      'groupMonthlyPayments',
-      'groupIncomeAdjustedDistribution',
-      'ourGroupProfile',
-      'groupSettings',
-      'ourUsername',
-      'userDisplayName'
-    ]),
-    needsIncome () {
-      return this.ourGroupProfile.incomeDetailsType === 'incomeAmount'
-    },
-    symbolWithCode () {
-      return currencies[this.groupSettings.mincomeCurrency].symbolWithCode
-    },
-    symbol () {
-      return currencies[this.groupSettings.mincomeCurrency].symbol
-    }
-
-  }
+  computed: {}
 }
 </script>
 
@@ -85,185 +55,31 @@ export default {
 .c-payments {
   margin-top: 1rem;
 
-  th, td {
+  th,
+  td {
     &:first-child {
       width: 41%;
     }
 
     &:last-child {
-      width: 10%;
-      padding-left: 1rem;
+      width: 5%;
 
       @include tablet {
-        width: 22%;
+        width: 20%;
       }
     }
   }
-
-  .c-header-checkbox {
-    line-height: 23px;
-  }
-
-  &-date {
-    @include phone {
-      display: none;
-    }
-  }
-
-  &-amount {
-    @include phone {
-      text-align: right;
-    }
-  }
 }
 
-.is-editing {
-  th, td {
-    &:first-child {
-      width: 10%;
-      @include phone {
-        width: 3rem;
-      }
-    }
-
-    &:last-child {
-      width: 40%;
-      min-width: 9.375rem;
-      display: table-cell;
-    }
-  }
-
-  .c-actions {
-    justify-content: flex-end;
-  }
-
-  .c-reset {
-    margin-left: 1rem;
-  }
-
-  .c-payments-date {
-    @include phone {
-      display: block;
-    }
-  }
-}
-
-.checkbox {
-  margin-right: 0.5rem;
-}
-
-.c-actions {
-  justify-content: space-between;
-}
-
-.c-actions,
-.c-user {
-  display: flex;
-  align-items: center;
-}
-
-.c-avatar {
-  margin-right: 0.5rem;
-}
-
-.c-avatar,
-.c-actions-month {
+.c-th-amount {
   @include phone {
-    display: none;
+    text-align: right;
   }
 }
 
-.c-actions-month {
-  margin-left: 0;
-  white-space: nowrap;
-}
-
-.c-actions-menu {
-  margin-left: 1rem;
-}
-
-.c-name {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.c-user-month {
-  display: none;
+.c-th-actions {
   @include phone {
-    display: inline-block;
-    margin-left: 0;
-  }
-}
-
-.c-tooltip-warning {
-  outline: none;
-
-  .c-tip{
-    background: $warning_0;
-  }
-
-  &:focus .c-tip {
-    box-shadow: 0px 0px 4px $primary_0;
-  }
-}
-
-// TODO: check if we need generic rule for this (keep until we know)
-// .c-tooltip {
-//   @each $name in $colors {
-//     &-#{$name} .c-tip{
-//       background: var(--#{$name}_0);
-//     }
-//   }
-// }
-
-.c-payments-amount-text {
-  @include tablet {
-    margin-right: 0.5rem;
-  }
-}
-
-.c-payments-amount {
-  &-info {
-    display: inline-flex;
-    align-items: center;
-
-    @include phone {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-end;
-    }
-  }
-
-  &-text {
-    @include phone {
-      display: block;
-    }
-  }
-
-  .is-small {
-    margin: 0;
-  }
-}
-
-.inputgroup .input {
-  @include phone {
-    padding-right: 2rem;
-  }
-}
-
-// Tooltip
-.c-tip {
-  width: 0.875rem;
-  height: 0.875rem;
-  border-radius: 50%;
-  margin-left: 0.5rem;
-  margin-right: 0.25rem;
-  font-size: 0.5rem;
-
-  &,
-  &:hover {
-    color: $background_0;
+    opacity: 0;
   }
 }
 </style>
