+++ frontend/views/containers/payments/RecordPaymentsList.vue
@@ -2,13 +2,14 @@
 table.table.table-in-card.c-payments.is-editing
   thead
     tr
-      th
+      th.c-th-checkbox
         label.checkbox
           input.input(type='checkbox' v-model='tableChecked')
-          .c-header-checkbox
+          span
+            i18n.sr-only Mark sent to all
       i18n(tag='th') Sent to
-      th.c-payments-amount
-      i18n.c-payments-date(tag='th') Amount sent
+      i18n.sr-only(tag='th') Due
+      i18n.c-th-amount(tag='th') Amount sent
 
   tbody
     payment-row-record(
@@ -22,8 +23,7 @@ table.table.table-in-card.c-payments.is-editing
 import { mapGetters } from 'vuex'
 import AvatarUser from '@components/AvatarUser.vue'
 import Tooltip from '@components/Tooltip.vue'
-import PaymentRowRecord from '@containers/payments/PaymentRowRecord.vue'
-import { humanDate } from '@view-utils/humanDate.js'
+import PaymentRowRecord from './payment-row/PaymentRowRecord.vue'
 
 export default {
   name: 'RecordPaymentsList',
@@ -61,15 +61,6 @@ export default {
         payment.checked = newVal
       }
     }
-  },
-  methods: {
-    // TEMP
-    async reset () {
-      console.log('Todo: Implement reset payment')
-    },
-    dueDate (datems) {
-      return humanDate(datems, { month: 'short', day: 'numeric' })
-    }
   }
 }
 </script>
@@ -77,188 +68,41 @@ export default {
 <style lang="scss" scoped>
 @import "@assets/style/_variables.scss";
 
-.c-payments {
+.c-payments.is-editing {
   margin-top: 1rem;
 
-  th, td {
-    &:first-child {
-      width: 41%;
-    }
-
-    &:last-child {
-      width: 10%;
-      padding-left: 1rem;
-
-      @include tablet {
-        width: 22%;
-      }
-    }
-  }
-
-  .c-header-checkbox {
-    line-height: 23px;
-  }
-
-  &-date {
-    @include phone {
-      display: none;
-    }
-  }
-
-  &-amount {
-    @include phone {
-      text-align: right;
-    }
-  }
-}
-
-.is-editing {
-  th, td {
+  th,
+  td {
     &:first-child {
       width: 10%;
+
       @include phone {
         width: 3rem;
       }
     }
 
+    &:nth-child(2) { // Sent to
+      max-width: 20vh;
+    }
+
     &:last-child {
-      width: 40%;
+      width: 35%;
       min-width: 9.375rem;
       display: table-cell;
     }
   }
-
-  .c-actions {
-    justify-content: flex-end;
-  }
-
-  .c-reset {
-    margin-left: 1rem;
-  }
-
-  .c-payments-date {
-    @include phone {
-      display: block;
-    }
-  }
 }
 
-.checkbox {
-  margin-right: 0.5rem;
+::v-deep .cpr-actions { // PaymentRow.vue
+  margin-right: 1rem;
+  justify-content: flex-end;
 }
 
-.c-actions {
-  justify-content: space-between;
-}
+.c-th-checkbox {
+  line-height: 1.5rem;
 
-.c-actions,
-.c-user {
-  display: flex;
-  align-items: center;
-}
-
-.c-avatar {
-  margin-right: 0.5rem;
-}
-
-.c-avatar,
-.c-actions-month {
-  @include phone {
-    display: none;
-  }
-}
-
-.c-actions-month {
-  margin-left: 0;
-  white-space: nowrap;
-}
-
-.c-actions-menu {
-  margin-left: 1rem;
-}
-
-.c-name {
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.c-user-month {
-  display: none;
-  @include phone {
-    display: inline-block;
-    margin-left: 0;
-  }
-}
-
-.c-tooltip-warning {
-  outline: none;
-
-  .c-tip{
-    background: $warning_0;
-  }
-
-  &:focus .c-tip {
-    box-shadow: 0px 0px 4px $primary_0;
-  }
-}
-
-// TODO: check if we need generic rule for this (keep until we know)
-// .c-tooltip {
-//   @each $name in $colors {
-//     &-#{$name} .c-tip{
-//       background: var(--#{$name}_0);
-//     }
-//   }
-// }
-
-.c-payments-amount-text {
-  @include tablet {
+  .checkbox {
     margin-right: 0.5rem;
-  }
-}
-
-.c-payments-amount {
-  &-info {
-    display: inline-flex;
-    align-items: center;
-
-    @include phone {
-      display: flex;
-      flex-direction: row;
-      justify-content: flex-end;
-    }
-  }
-
-  &-text {
-    @include phone {
-      display: block;
-    }
-  }
-
-  .is-small {
-    margin: 0;
-  }
-}
-
-.inputgroup .input {
-  @include phone {
-    padding-right: 2rem;
-  }
-}
-
-// Tooltip
-.c-tip {
-  width: 0.875rem;
-  height: 0.875rem;
-  border-radius: 50%;
-  margin-left: 0.5rem;
-  margin-right: 0.25rem;
-  font-size: 0.5rem;
-
-  &,
-  &:hover {
-    color: #fff;
   }
 }
 </style>
